# backend_fashion_app/internal/common/httpx/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.AuthResponse:
    properties:
      refreshToken:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/api.UserInfo'
    type: object
...
  api.LoginRequest:
    properties:
      credential:
        example: test@example.com/0912345678
        type: string
      password:
        example: Cong123@
        type: string
    required:
    - credential
    - password
    type: object
...
  api.RefreshRequest:
    properties:
      refreshToken:
        type: string
    type: object
  api.RegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        example: alice@example.com
        format: email
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - confirmPassword
    - email
    - password
    - phone_number
    type: object
...
  api.UserInfo:
    properties:
      id:
        type: integer
      role:
        type: string
    type: object
...
  resp.Envelope:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Auth (local, Google) + Profile (customer/shop) + Catalog
  title: MyFashion API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập local (email/password)
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.AuthResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Thu hồi tất cả refresh token và blacklist access token hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Logout
      tags:
      - Auth
  /auth/me:
    get:
      description: Lấy info hiện tại từ JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Me
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Đổi refresh token lấy access token mới (có xoay vòng và phát hiện
        tái sử dụng)
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Đăng ký tài khoản customer (chỉ customer, shop được cấp thủ công)
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/api.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
      summary: Register
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"